Removed linked IDP's from SP.
We can not determine in a non-DDOS way for Service Registry what IDPs
are connected to what SP's. Until we figure out how to do this remove
this functionality from the offered-services.json. The functionality we
need is provided by the Prodiver.isLinked method. This method is set in 2
different ways:

1. Via janusClient.isConnectionAllowed(spEntityId, idpEntityId)
2. Via cross-referencing the SpList en de AllowedSpList(idpId)

It is not certain this results in the same answer and is unwanted
duplication of businesslogic.